requirements.txt
Django>=4.1,<4.2
djangorestframework>=3.14,<3.15
drf-spectacular>=0.25<0.26
PyMySQL==1.0.2
djangorestframework-simplejwt>=5.2.2<5.3
